+++ src/hcl_model/model_hcl_generic.py
@@ -90,9 +90,10 @@ class HandCraftedLinearModel(TimeSeriesModelArchetype):
         num_simulations: int = None,
         **kwargs
     ) -> pd.DataFrame:
-        self._exog = self._prepare_exog(exog=X)
+        if X is not None:
+            self._exog = pd.concat([self._exog, X])
         nobs = self._get_num_observations(self._endog)
-        self._check_exogenous(exog=X, nobs=nobs, num_steps=num_steps)
+        self._check_exogenous(exog=self._exog, nobs=nobs, num_steps=num_steps)
         endog_updated = pd.concat(
             [self._endog, pd.Series(np.empty(num_steps), name=self._get_endog_name(), index=self._exog.index[nobs:])]
         )
@@ -121,12 +122,13 @@ class HandCraftedLinearModel(TimeSeriesModelArchetype):
         self,
         num_steps: int,
         num_simulations: int,
-        endog: pd.Series = None,
-        exog: pd.DataFrame = None,
+        y: pd.Series = None,
+        X: pd.DataFrame = None,
         weights: Union[Sequence, float] = 1.0,
         **kwargs
     ) -> pd.DataFrame:
-        self._endog, self._exog = self._prepare_data(endog=endog, exog=exog)
+        if X is not None:
+            self._exog = pd.concat([self._exog, X])
         nobs = self._get_num_observations(self._endog)
         self._check_exogenous(exog=self._exog, nobs=nobs, num_steps=num_steps)
         num_params = self._get_parameters().shape[0]